+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.4.0] - 2021-01-29
+
+### Added
+
+- `--skip-compressed` - skip compressed files if they already exist
+
+### Changed
+
+- compression extensions like `gz`, `zz`, `br`, etc. excluded by default
+
 ## [4.3.0] - 2020-10-30
 
 ### Added